Move migration engine integration tests into the crate
This allows us to delete lib.rs and avoid compiling the test harness
for every single suite. The crate compiles faster, without any cost in
terms of size, dependencies or compile time to the release binary.

That way, if we decide to expose the migration engine as a library some
day, we can have a more thoughtful public API.