findOne to findMany optimization
This feature implements an optimization to a special case in batch
operations. To trigger the optimization, the query needs to follow
certain rules.

An example batch request that can be optimized:

```json
{
    "batch": [
        {
            "query":"query {findOneUser(where:{firstName:\"Musti\"}) {firstName lastName}}",
            "variables":{}
        },
        {
            "query":"query {findOneUser(where:{firstName:\"Naukio\"}) {firstName lastName}}",
            "variables":{}
        }
    ]
}
```

Here we have two findOne queries with a different query param but the
same selection set. This will result a compacted query where we look for
all the users with a single query, then map the data back to the
separate requests.

An example request that will not be optimized:

```json
{
    "batch": [
        {
            "query":"query {findOneUser(where:{firstName:\"Musti\"}) {lastName}}",
            "variables":{}
        },
        {
            "query":"query {findOneUser(where:{firstName:\"Naukio\"}) {firstName lastName}}",
            "variables":{}
        }
    ]
}
```

In this case the selection set is different between the queries and the
system will spawn separate selects to fetch the data, skipping the optimization.

A query with a compound search should be optimized with a bit different
resulting SQL:

```json
{
    "batch": [
        {
            "query":"query {findOneUser(where:{firstName_lastName:{firstName:\"Musti\",lastName:\"Naukio\"}}) {firstName lastName age email}}",
            "variables":{}
        },
        {
            "query":"query {findOneUser(where:{firstName_lastName:{lastName:\"Musti\",firstName:\"Naukio\"}}) {firstName lastName age email}}",
            "variables":{}
        }
    ]
}
```

The implementation of the optimization is not in the sweet spot and a
recommendation would be to move the optimization altogether to the
javascript layer.