Drop the foreign keys of dropped tables separately
Cyclic foreign key dependencies would error when dropping the table at
either end of the foreign key. Since we run `ALTER TABLE`s before `DROP
TABLE`s, we drop the foreign keys in and to dropped tables before
dropping the tables, which should solve the problem.

SQLite is a special case because it does not have a way to drop foreign
keys separately from the table. We disable foreign key checks and rely
on the schema to forbid relations pointing to non-existent tables.