Restructure SqlDestructiveChangesChecker
The purpose of this commit is to separate pure checks (what steps in the
migration _could_ cause problems) from IO. This serves multiple
purposes:

- It makes it easier to reuse the results of previous checks, e.g. to
avoid checking multiple times that there are non-null values in some
column.
- It makes it easier to provide alternative warnings when the database
we want to migrate is not accessible, or the checks time out. (e.g.
"This migration can be destructive if there are non-null values in such
and such column"). In other terms, we can do destructive change checks
without IO.
- It makes the code clearer, because checks have to be explicit about
what information they need about the current state of the database.
- It is a better default: the checks do not rely on information fetched
from the database at the last moment, they just use it to refine their
warnings or be more optimistic (e.g. it's safe to drop a table if it is
empty).

See the diff for a more in-depth explanation in doc comments.