Make database steps backwards-compatibility irrelevant in listMigrations
Here we only deserialize database migrations to render them for the CLI. The only change here is that we still try, but won't crash if the format changed, and only return an empty array of steps instead. @Jolg42 I did my best to not break anything, but I'm not 100% that this won't cause light breakage in the CLI.

If we still want to render useful information in presence of breakage in the future, we could pre-render and store migration steps rendered to strings (SQL queries or descriptive string). I did not do this because it is unclear what will happen with this feature.