Stop relying on persisted JSON database migration in unapplyMigration
The goal in this PR is to avoid user-facing breakage when we change the shape of database migrations (at the moment this means SqlMigration), since we will most likely break things multiple times as we start working on migrations again and add mssql support. This is also decoupling breaking changes in the migration engine and sql-schema-describer.

We were deserializing and using database migrations from CLI input/the migrations table in two places.

listMigrations
Here we only deserialize database migrations to render them for the CLI. The only change here is that we still try, but won't crash if the format changed, and only return an empty array of steps instead. @Jolg42 I did my best to not break anything, but I'm not 100% that this won't cause light breakage in the CLI.

If we still want to render useful information in presence of breakage in the future, we could pre-render and store migration steps rendered to strings (SQL queries or descriptive string). I did not do this because it is unclear what will happen with this feature.

unapplyMigration
The way unapplyMigration works before this PR:

Grab the last migration from the migrations table, and use the rollback database migration from the stored database migration.
This has two issues:

If the database steps format breaks, we won't be able to rollback
The current state of the database is ignored, so the steps could easily be wrong
This PR changes unapplyMigration to:

Get the last two migrations from the migrations table, so we have the current datamodel and the one before that (otherwise assume an empty datamodel).
Diff these
Apply the migration from last datamodel to the one before that like any regular migration
The only difference between applyMigration and unapplyMigration now is how they interact with the migrations table, and also watch mode. This means some of the unapply infrastructure is not needed anymore, but I haven't outright removed it yet.