Remove ForeignKeyDefaultValue warning from migration-engine
This check was necessary with the old relations syntax, because default
values on relation fields could not be expressed, so we would always
drop the defaults on introspected schemas. With the new relations syntax
the defaults can be expressed, so we can assume the default was removed
by the user and this warning is not needed anymore.

Closes https://github.com/prisma/prisma-engines/issues/824