Initial implementation of schemaPush command
This is something we want to experiment with. It takes a prisma schema and migrates the database to match it, without interacting with the migrations, which we think should be a separate workflow. This is meant as a dev workflow, when you are doing initial iteration on your application and do not care about migrations yet.

The name of the command is a draft name, it will likely change.

Open work items: we should persist schema pushes to the migrations table for auditability/debuggability.