Preserve arity of altered enum columns on MySQL
Problem: AlterEnum steps on MySQL would only attempt to change the type of the column, but lost the arity in the process.

Two solutions were envisaged:

Change AlterEnum rendering to share most of the architecture of ColumnChange::AlterColumn, so it would produce a valid column migration.
Since enum changes are implemented by a MODIFY on the column, absorb enum diffing into column diffing on MySQL.
This PR went with (2), because (1) has the big downside of producing two steps for a change that can be done in a single statement, and it sticks closer to the way enums work on MySQL — that way it will not break as easily if we change how we represent MySQL enums in sql-schema-describer.

closes #1032