Let native column types be created in MySQL migrations
Last part of the first phase of https://github.com/prisma/prisma-engines/issues/72

After this PR, the migration engine will create tables and columns with native types on MySQL (postgres is already implemented). The next step will be diffing existing columns with native types, so we can migrate between them. I expect that second step to be significantly harder.

This unblocks work on native types in the query engine.