Change the way how we convert Decimal to f64
In our JSON responses we go from `Decimal` to `f64`. rust_decimal
recently changed their conversion algorithm from `Decimal` -> `String`
-> `f64` to a faster mantissa algorithm that goes `Decimal` -> `f64`
directly. This seems to cause some loss of precision for now, so we'll
do the conversion as the crate used to do through a `String`, that is
slower but keeps the accuracy.