Make KnownError::new() infallible
serde_json::to_value can technically panic, but 1. it is very rare, 2.
it would always be caught in tests in our case, so it's more ergonomic
to assume it will not fail. We do not handle that error properly anyway.

An alternative to a hard error when failing to render would be rendering
to `null` and logging the error.