Include more context in SqlMigration::RedefineTables
This is for redefining tables on sqlite. This commit makes it a step
that is inferred like all the others.

The idea is to refactor towards making the differ structs private to
thee sql_schema_differ module.

---

Start refactoring towards making differ structs private

The SqlMigration struct should hold all the necessary context.