Migrations on SQL Server (#1372)
* Migrations on Microsoft SQL Server

* Make the shadow database cleanup more robust

* Primary key recreation flag for column recreate

For SQL Server, we want to recreate the primary key before recreating a
column that's part of it. We do not want this behavior on other
databases, thus adding a new flag to the flavour.

* Make alter table to work as a module

Much nicer and following the crate standards this way, without needing
to add a completely unnecessary constructor creation...

* Hide stuff in SqlSchema debug, change few fields

* Flavour out the imperative migrations table

* Add bigger timeouts on tests

* Warn about NotCastable column changes

* Add back assertions

Also fixes one test to not re-create the whole table (to spawn warnings
on SQL Server). The barrel `types::primary` will make the column
autocomplete and the next migration removes the identity, spawning a
recreation of the whole table.

* Fix non-castable warnings

* Cargo fmt

* Update base64 crate