N-API Support for Query Engine (#1544)
To compile:

```bash
> cargo build --release -p query-engine-napi
```

This will compile a file in `target/release/libquery_engine_napi.so`. To
be able to use the file in node, please move it elsewhere and rename as
`libquery_engine_napi.node`.

From node.js, you can then load the module, and initialize a new Query
Engine:

```javascript
const napi = require('./libquery_engine_napi.node');

let qe = new napi.QueryEngine(datamodel);
```

The `datamodel` in the parameter should be a fully loaded datamodel as a
string, including configuration, generator, schema and all that. The
rust code then parses and validates the model, loads the configuration,
loads the feature flags from the datamodel and gives back a new instance
of a `QueryEngine` class.

Because we don't have async constructors in javascript, the next step is
to connect the engine:

```javascript
await qe.connect();
```

And then execute delightful queries:

```javascript
let response = await qe.query(gqlJson);
```

The request and response format should follow the HTTP server. Please be
aware that some of the errors come back in the response (as before), but
all methods (excluding the raw query enabling) can also throw. If the
error happens in the layer above core, we throw a `GenericError` in
JavaScript, that has a json-encoded payload in the error message.