Disable shadow database url validation in datasources
Something unintuitive is happening in practice, where even URLs that are
not the same are transformed, and passed through validation as the same
URL, at least in certain circumstances. We should discuss this in the
meeting planned for wednesday.

There is a second layer of validation in the migration engine, we well
rely on it for now.