Foreign keys should always be in the written order
In this bug, we iterate first from the table's columns, finding the
columns to the given foreign key. If the table has columns in a
different order than in the key, we try to create a foreign key with the
compound columns in a wrong order.

We should instead iterate from the foreign key columns, which is
typically a bit faster but also preserves the column ordering.