Do not fail on missing/empty shadow database URLs
The current behaviour forces users to specify a shadow database URL in
their environment, even for situations (e.g. deployment) where you would
not need, and indeed should not use a shadow database.

This PR makes the datamodel parser ignore the shadow database URL if it
is empty, or the env var is not defined.

closes https://github.com/prisma/prisma/issues/5704