A channel-based log stream for N-API (#1665)
* A channel-based log stream for N-API

When creating the Query Engine, a new parameter is added to the
configuration:

```javascript
let qe = new QueryEngine({
  logLevel: "info",
  ...
});
```

This parameter can be any of: `trace`, `debug`, `info`, `warn`,
`error` or `off`. The parameter is case-insensitive and sets the
minimum level. `trace` logs everything, `off` logs none.

Additionally, the engine got a new method:

```javascript
let event = await qe.nextLogEvent();
```

This is paused until a log event is produced to the channel, which is
returned from the method. The event is stringified JSON, which can be
parsed on the JavaScript side. The method can only be called when the
engine is connected.

The log level in the returned event is always capitalized.

The log channel is bounded and has a maximum capacity of 10 000 events.
If the channel is not consumed fast enough and it gets full, further log
events will get dropped and an error gets produced to STDERR.

If the method returns a `null` value, it means engine got disconnected,
and a new callback should get registered when re-connecting the engine.

Please be aware of setting the `LOG_QUERIES` environment variable to any
value, if wanting to get the executed SQL queries on the channel. The
queries will then arrive with the log level of `INFO`.

* Remove panic with feature flag double-init, remove file and line from log events

* Add napi to `flags!`

* Special disconnect log event