fix: findUnique batch with date arguments (#1771)
The findUnique batch optimization expects inputs to be strictly equal to outputs. In the case of dates, they were serialized in a different format than the input format, causing nulls to be returned.

Given that clients always send input dates in the following format, `1999-05-01T00:00:00.000Z`, we updated the output date serialization format to always render dates in the same format as above.