Handle more errors when trying to create a db (#1827)
We try to connect to some available postgres database while creating a
new database. The order is `postgres`, `template1` and `defaultdb`. We
try again, if the database does not exist, but crashed and burned if we
didn't have access to the given database.

In case of the cloud version of timescaledb, we connect first to
`postgres` which does not exist, then `template1` which gives us access
denied. The third one would've be `defaultdb` that would work, but we
already exited during `template1`.