Skip tracing spans and records on napi logger
By creating a new `Map` for every span and record we visit on every span
or record has a quite dramatic effect on our RSS use. What we did was:

- Create a new `Map` for every span
- And create a `Map` again for every record

Originally, this was not causing any havoc due to us not having so many
spans in the system... Until the tracing instrumentation was added to
many of the functions. Suddenly we called `new_span` for every function
in the path from client to server. And we created a lot of `Map`s.

A lot of `Maps`...

So, we don't really use ANY of this info in the context of js logger,
we don't really need to decorate any of it in this part of the code.
Actually removing this code did NOTHING to what our client can log.

But, the previous hockey stick of memory usage -- which I stopped when we
climbed from 180MB to a gigabyte -- is not there anymore. After running
my test script with this change for 20 minutes, the RSS climbed from 180
to 210MB, sometimes dropping down a bit and then climbing back to the
previous value.