Quick and dirty parameter hints
SQL Server uses type information for `INSERT` queries with `OUTPUT` to
counteract the restriction with `OUTPUT` when using triggers in the
database. This creates a list variable that gathers the output data, and
this variable is typed.

We originally just gave hints and our best effort for the type, only to
find out that with native types, such as `decimal(26,0)`, this value
would not fit to the default `decimal(32,16)` and would error out.

Same when returning `nvarchar(max)` and so on. Although these are not
yet used in `OUTPUT`, but will be if we ever start using `RETURNING`
inserts for more than just the identifiers.