Remove ignore_datasource_url in datamodel parser
The problem solved by ignore_datasource_url is that sometimes, we want
to parse a schema without resolving the env vars it contains, for
example in `getConfig` in the CLI.

The problem with ignore_datasource_url is that it permeates through the
API, justifying multiple `parse_*` top-level functions for different
combinations of special cases in the datamodel crate.

The more general problem is that ideally, schema parsing should not
involve resolving env vars, and at the same time, we want these URLs
from the environment to be validated.

What this PR implements is the following:

- Remove ignore_datasource_url and the corresponding top-leve `parse_*`
  functions, and reduce the top-level API surface of the datamodel
  parser.
- Stop resolving env vars during parsing. The parsed datasource will
  still contain the unresolved env vars when the config is parsed.
- Ensure the contents of the env var are validated by making it private,
  and enforcing that it is read with `Datasource::load_url()`, which
  performs validation.