Make generator binaryTargets a StringFromEnvVar
Fixes a regression documented in
https://github.com/prisma/prisma/issues/7295 : the generatorConfig was
_implicitly_ potentially from the environment. Since we do not resolve
env vars during parsing anymore, using the env function in binaryTargets
started returning errors from DMMF generation.

The solution is to make binaryTarget members first-class
`StringFromEnvVar`s. Note that this is a small breaking change in the
shape of the DMMF that needs to be coordinated with the Typescript code.

closes https://github.com/prisma/prisma/issues/7295