Make configuration parsing internally infallible
From a code organization perspective, validation errors are not error
conditions for the parser. We do not want them to affect the control
flow, but on the contrary to accumulate them and continue working with
broken input as much as it makes sense. To go further than this, we will
have to separate validation from Datamodel instantiation, longer term.

Observation: the value validators returning results tends to produce
slightly verbose match statements all over the place, we might want to
look into that in the future.