Group parser state into `ParserDatabase`
* Make Names private to ParserDb

We want multiple components, like `Names`, to build up information
during the validation process. `Names` was implemented as an isolated
struct that makes sense in concert with / augments a SchemaAst. I
experimented with a component for field type checking yesterday, and
noticed that code organization is not easily extensible. It is easier to
have the state live in a single place where, for example, the type
checking component can take advantage of name resolution component.

For prior art, see rust-analyzer's layered Salsa databases pattern.  All
state is available to the root database, and every layer has access to
the layers below (starting with source text):
https://github.com/rust-analyzer/rust-analyzer/blob/38ae18b7592f97a7058d97928307bccbd881a582/crates/ide_db/src/lib.rs#L45.

Additionally, bundling the components with a reference to the AST lets
us expose a simpler API to the validation code (there is no extra
urgument for the AST in every function).