Correct detection of multiple cascading paths (#2192)
* Detect multiple cascading paths between models

On SQL Server it is not allowed to have more than one path between two
models if the referential actions of the relations are something else
than `NO ACTION`. This code detects these paths early, returning a
validation error for the user.

* Clippy

* Only check fields that hold from/references

Also break out of cycles in path detection