Compile and run all unit tests in CI
Instead of an allow-list of crates to test, use a deny-list of crates to
exclude. The test crates we exclude are those that need a database
connection to run.

We already had an example of forgotten crate that stopped compiling and
that did not make CI fail, this should prevent it.