Sanitize the env param on Query Engine init
When the node.js process sends down the `process.env` as a parameter, we
cannot guarantee the values are all strings. Usually they are, but we
have some indication this can be overwritten by libraries such as server
frameworks. Here we expect the data to be anything JSON can hold,
sanitizing it to a hashmap of string values before using it in the
system.