Move libs/prisma-models to query-engine/prisma-models
A long time ago, logic from prisma-models was used in the migration
engine, but it hasn't been the case in a long time. To clarify that
prisma-models is the QE representation of a datamodel _and the
corresponding QE API_, let's move the crate to the query-engine
directory.

This was discussed and agreed on in a recent Rust group meeting.