Implement a first version of `db push` on MongoDB
The layout of the mongodb migration connector mimics the sql migration
connector: schema calculator, differ, migration, etc. It follows the
same MigrationConnector API and general flow.

It also includes introspection (in client_wrapper.rs) and a schema
representation (in schema.rs), similar to sql_schema_describer.

The test setup is different and explained in
`tests/migration_tests/migrations.rs`.

This PR also introduces a mongo client wrapper to the
mongodb-migration-connector. It's a learning from
sql-migration-connector that we want to fully own the database query API
in the connector.