Comment out [lib] section of prisma-fmt/Cargo.toml
This is because builds are failing on linux-musl (Alpine), as dynamic
linking is just not supported. That section is for the wasm builds,
which are not in CI yet, so it should improve the situation.

The wasm build situation is less than ideal, I want to explore
alternatives, and in particular not building from prisma-engines.