Incorporate `create_database()` and `drop_datbase()` into GenericApi
`create_database()` is the implementation behind `migration-engine
--create-database`, used by the Prisma CLI. From the very beginning of
the Rust migration engine until this commit, it was implemented as a
separate top-level free-standing function that does not go through the
migration engine JSON-RPC API.

The reason behind this was that the JSON-RPC API eagerly connected to
the database, which tends to not work well for `create_database()` for
obvious reasons.

The consequence was a larger top-level library API for the migration
engine, and a different way to call the migration engine for the Prisma
CLI. This is unnecessary complexity.

Here, we are starting to reap the benefits of
b5a631aa090a4e69da6f48d2b1bb459dbe317a4b by unifying the migration
engine API: for rust code, only `GenericApi`, and for the CLI, only
JSON-RPC. The previous API is kept as a shim on top of the new one.