Add special handling for mongodb+srv URLs in migration engine CLI
`datasource_from_database_str()` should not exist, there is no good reason
it does. It is a consequence of old tech debt in the typescript CLI.

There is no test because we would need a new mongo test setup for the
CLI, and we are in a rush (release day).