Make datamodel providers' native type constructors constant
The justification for this is not performance (although it almost
certainly be better), but that we want to enforce that there is no
dynamic behaviour in datamodel connectors: for a given provider, the set
of available native types should always be the same. The 'static
lifetimes enforce that.

The next step should be capabilities and constraint violation scopes.