Make capabilities and referential integrity const in datamodel connector
Similar to the two previous pull requests for constraint scopes and
native type constructors, the datamodel connector is used in parsing:
the parsing rules should be consistent. Making the capabilities const
prevents changing them at runtime, which wouldn't be desirable for
schema validation.

A connector should not dynamically change capabilities. Version
differences and settings based on the prisma schema do not belong in the
datamodel connector — at least not yet and not in a dynamic way.