datamodel: Move AST renderer to schema-ast crate
This finishes the easy parts of the move, and is a requirement if we
want to move the parser database out of core.

The last AST module that lives in core is the reformatter, but we can't
move it easily because it depends on the DML, which is a dependency we
can't take on in schema-ast.