Prototype prisma-fmt driven LSP completion
This adds a new public method to prisma-fmt, and the infrastructure to
provide completion at a given point in a Prisma schema. As part of this,
we must make the AST parser more resilient to broken schemas. In this
PR, we do it by allowing _empty arguments_ in attributes:

```
@relation(onDelete: )
```

This now parses to an AST within an `empty_attributes` vector on the
attribute as a result. We validate against it in validation. This is
more code than before, but it has two advantages:

- Instead of getting a vague syntax error, you get a better, more
  specific validation error.
- You don't need to already have a completely valid AST before you get
  completion suggestions.

The idea would be to expand what the AST recognises in this direction
over time. The quality of our error messages would also see improvements
as a side-effect.