Abstract away how the ME calls to the CLI
In the CLI -> ME direction, we use JSON-RPC, which is well specified.

In the ME -> CLI direction, we use a custom protocol over the JSON logs,
which is a hack. What we want is JSON-RPC in both directions. In order
to do this, we use dependency injection: a migration connector can
receive a `ConnectorHost`, that exposes client-side IO as an API. In
this commit, we implement it for the old protocol, but it will migrate
to JSON-RPC in a follow up.