Isolate qe-setup from the rest of migration engine
Now the qe test setup only relies on the public API of migration_core.
All the extra details are moved from the migration connectors to the new
qe-setup crate. Future direction would be doing more in the QE itself,
to avoid creating ME connections for each test for example, which would
probably improve the duration of test runs a lot.

Extraneous change in this commit: make migration_core::migration_api()
sync. There's no reason it can't be anymore, and that simplifies things.