prisma-fmt: Fix cursor position resolution with CRLF
text_document_completion must convert from an LSP `Position` (two
integers: line and character) to a byte offset in the schema, since this
is what our AST Spans are. Until this commit, it counted line endings
for one character, but they can be two characters on windows.

The solution that implemented and tested here is to still treat '\n' as
the line delimiter, but without skipping carriage returns. We achieve
this by not using the `str::lines()` iterator from std.