Commenting out wonky fields in Mongo introspection (#2651)
* Commenting out wonky fields in Mongo introspection

In a case where a model or a type has a field we cannot represent in Prisma, and
we cannot rename to anything, we give up commenting the field out and asking the
user to rename it manually.

Also prevents using a datamodel with a composite type that has no fields.

* Warn if commented out fields

* Commenting out wonky fields in Mongo introspection

In a case where a model or a type has a field we cannot represent in Prisma, and
we cannot rename to anything, we give up commenting the field out and asking the
user to rename it manually.

Also prevents using a datamodel with a composite type that has no fields.