Delete diagnostics::ConnectorError
This type was part of datamodel connector a long time ago, and meant to
fulfill the need for a type of errors that couldn't depend on AST
information, that would be grafted on after the fact in datamodel core.
Now that we have ParserDatabase, this is pure accidental complexity
anymore, we know where we are in the AST.

This commit deletes ConnectorError and replaces it with regular
datamodel errors. It makes validation code both in datamodel-connector
and in the core more straightforward.