me: Implement --exit-code param
The `diff` RPC method now takes an optional `exitCode` boolean param. If
it is set to true, the `exitCode` field in the `DiffResult` returned by
the method will be set to 0 if there is no detected diff (the schemas
are in sync), and 2 otherwise.

Issue: https://github.com/prisma/prisma/issues/11566