datamodel: Export dml from a single module
All crates outside of libs/datamodel depend on datamodel/core, it is the
public API of the datamodel functionality.

The situation before this commit is that datamodel core reexports dml
types from both the `dml` module and at its top level, making the
top-level API surface very large and all over the place.

This commit makes it export the types of the dml data structure only
from the dml module, which makes code across engines more uniform,
consistent and predictable.