Start separating reformatting logic from lifting logic
So far, the reformatting logic consisted in comparing the actual AST
with the datamodel lifted from it, and seeing what's missing. This is
problematic because it couples reformatting and the internal
representation of the datamodel in the QE, making the dml more ast-like
than it needs to be, while at the same time not giving us enough control
on how reformatting actually works.

This commit starts making the reformatting logic local to the
reformatter, with three expected outcomes:

- Decoupling from dml, as motivated in the first paragraph above
- More intelligible logic
- Logic that we can reuse for code actions if and when we move from
  adding fields in reformatting to code actions. Since code actions will
  be based on parser database, dml wouldn't be an option there.