me: Make migrations not depend on default_int_size param
The gist of this commit is that we have Int and Bigint in the PSL. On all connectors, Int means 32 bit signed integer, and BigInt means 64 bit signed integer.

On Postgres, we map Int to INTEGER in a create table, and that's good and accurate.

On cockroachdb before this commit, we map Int to INTEGER, but the size of INTEGER depends on the default_int_size configuration (session or server) param, and by default it's 8. This commit goes around that by mapping int to INT4 to reflect Prisma expectations. BigInt is still mapped to BIGINTEGER, which is correct.

We use the cockroach-specific `INT4` to avoid integer sizing
ambiguities. For the complete solution, we will have to deal with the
different type of autoincrementing primary keys, because two out of the
there force INT8.

Addresses the default_int_size part of https://github.com/prisma/prisma/issues/12113