Mongo logging (#2752)
Support for query logging for MongoDB queries, we use the same format used by the Mongo Shell and Mongoose, so the query could be executed directly in the Mongo shell. A few important things need to be noted:

In DEBUG builds, the queries would be indented for easy reading
In RELEASE builds, the queries are plain, without indentation or car returns
This works exactly like the existing SQL query logging but a few details are not included:

Query execution time (maybe in a near future)
Query parameters (we don't use it)
When we retake logging this would be an area where things could be heavily simplified using spans for query logging. Another important improvement would be using a logging macro instead of the current functions.

This closes prisma/prisma#11419