fix(qe) Add dispatcher to actors (#2818)
This fixes the issue that when itx are used with the node-api
that no logs are collected for any of the transactions. This is because
we don't have a global subscriber. We fix this by getting the
current dispatcher and wrapping the new spawned tasks with it. All logs
will now go to that dispatcher.